rename develop folder to models
