+++ oct/pages/__init__.py
@@ -1,2 +1,2 @@
-from .base.base import Success as Success
-from .develop.return_page import ReturnPage as ReturnPage
+from .base.base import Success
+from .models.return_page import ReturnPage
